% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansi.R
\name{ansi}
\alias{ansi}
\title{Graphics device for ANSI output}
\usage{
ansi(
  filename = NULL,
  width = NULL,
  height = NULL,
  colour_depth = 8,
  plain_ascii = FALSE,
  pow = 1,
  char_lookup_table = 1,
  font_aspect = 0.45,
  ...
)
}
\arguments{
\item{filename}{If given, write ANSII to this file, otherwise write to console.}

\item{width, height}{dimensions of text output (in characters). Default: NULL
(auto-detect)}

\item{colour_depth}{ANSI colour depth. Default: 8.  8 is best for most terminals
(including the Rstudio console).   24-bit depth is not supported by
all terminals - use with caution.}

\item{plain_ascii}{Print ASCII representation only. default: FALSE}

\item{pow}{only valid when \code{plain_ascii = TRUE}. raise intensity to this
power before conversion to an ASCII character.  Default: 1}

\item{char_lookup_table}{only valid when \code{plain_ascii = TRUE}. Choose from
the available character mappings. Default: 1. Possible values 1, 2 or 3}

\item{font_aspect}{Character spacing horizontally and vertically are almost
never the same. On many terminals character resolution vertically is
half the resolution horizontally.  Adjust this value if circles don't
look right. Default: 0.45}

\item{...}{other parameters passed to the rdevice}
}
\description{
Graphics primitives will be rendered using ANSI characters.
}
\details{
Uses \code{devout::rdevice()}.
}
